Fix some release build issues
1. Currently the release build of CIV_R cannot boot up. Apex
files cannot be resigned successfully due to the restriction
of PATH tools. The tool of debugfs cannot be found in the
default PATH variable, need to append SOONG_HOST_OUT to the
PATH for sign_target_files_apks command.

2. There is no release build for the GPT image currently.
However, PGP ships a GPT image in its Release.deb instead
of a flashfiles package. Need to create a release build
for the GPT image from a flashfiles release package.

Tracked-On: OAM-100098
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>